Rename module SofaMisc with SofaMiscOther
