--- modules/SofaMisc/SofaMiscConfig.cmake.in
@@ -1,12 +1,0 @@
-# CMake package configuration file for the SofaMisc module
-@PACKAGE_INIT@
-
-find_package(SofaMiscTopology REQUIRED)
-
-### Is the target existing ?
-if(NOT TARGET SofaMisc)
-    ### If not we are building out of tree.
-    include("${CMAKE_CURRENT_LIST_DIR}/SofaMiscTargets.cmake")
-endif()
-
-check_required_components(SofaMisc)
