+++ modules/SofaMiscOther/SofaMiscOtherConfig.cmake.in
@@ -1,0 +1,13 @@
+# CMake package configuration file for the SofaMiscOther module
+@PACKAGE_INIT@
+
+find_package(SofaMiscTopology REQUIRED)
+find_package(SofaGeneral REQUIRED)
+
+### Is the target existing ?
+if(NOT TARGET SofaMiscOther)
+    ### If not we are building out of tree.
+    include("${CMAKE_CURRENT_LIST_DIR}/SofaMiscOtherTargets.cmake")
+endif()
+
+check_required_components(SofaMiscOther)
