+++ modules/SofaMiscOther/src/SofaMiscOther/MeshTetraStuffing.cpp
@@ -19,7 +19,7 @@
 *                                                                             *
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
-#include <SofaMisc/MeshTetraStuffing.h>
+#include <SofaMiscOther/MeshTetraStuffing.h>
 #include <sofa/core/visual/VisualParams.h>
 #include <sofa/core/ObjectFactory.h>
 #include <sofa/simulation/Simulation.h>
