+++ modules/SofaMiscOther/src/SofaMiscOther/config.h.in
@@ -23,10 +23,10 @@
 
 #include <sofa/config/sharedlibrary_defines.h>
 
-#ifdef SOFA_BUILD_MISC
-#define SOFA_TARGET @PROJECT_NAME@
-#  define SOFA_SOFAMISC_API SOFA_EXPORT_DYNAMIC_LIBRARY
+#ifdef SOFA_BUILD_MISC_OTHER
+#  define SOFA_TARGET @PROJECT_NAME@
+#  define SOFA_SOFAMISCOTHER_API SOFA_EXPORT_DYNAMIC_LIBRARY
 #else
-#  define SOFA_SOFAMISC_API SOFA_IMPORT_DYNAMIC_LIBRARY
+#  define SOFA_SOFAMISCOTHER_API SOFA_IMPORT_DYNAMIC_LIBRARY
 #endif
